fix(global): fixing all

+++ src/app/shared/modal-delete/modal-delete.component.ts
@@ -1,6 +1,7 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, Input, OnInit } from '@angular/core';
 import { SectorService } from 'src/app/core/services/sector/sector.service';
 import { StickerService } from 'src/app/core/services/sticker/sticker.service';
+import { SubsetService } from 'src/app/core/services/subset/subset.service';
 
 @Component({
   selector: 'app-modal-delete',
@@ -9,20 +10,34 @@ import { StickerService } from 'src/app/core/services/sticker/sticker.service';
 })
 export class ModalDeleteComponent implements OnInit {
 
+  @Input() sectorName:string;
+  @Input() subsetName:string;
+  @Input() opcion$:string;
+
   public opcion: string = '2';
+
+
   constructor(
-    public sectorService: SectorService,
-    public stickerService: StickerService
-  ) { }
+    private sectorService: SectorService,
+    private stickerService: StickerService,
+    private subsetService: SubsetService
+  ) {}
 
   ngOnInit(): void {
   }
   setOption(num: string) {
     this.opcion = num;
   }
-  buttonDeletePressed(){
-    this.stickerService.deleteAllStickersOfSector();
-    this.sectorService.deleteSector.emit();
+  buttonDeletePressed(selected:string){
+    if(selected == this.sectorName){
+      this.stickerService.deleteAllStickersOfSector();
+      this.sectorService.deleteSector.emit();
+    }
+    else{
+      if(selected == this.subsetName){
+        this.subsetService.deleteSubset.emit();
+      }
+    }
   }
 
 }
